Changes on Pre-Processing data

+++ data/data_handler.py
@@ -16,11 +16,11 @@ class DataHandler(object):
         directory_input = os.path.join(os.path.dirname(__file__), 'input')
         directory_log = os.path.join(os.path.dirname(__file__), 'log', 'input')
 
-        #removes all files from the input directory
+        #removes all network files from the input directory
         for file in os.listdir(directory_input):
-            
             filename = os.fsdecode(file)
-            os.remove(os.path.join(directory_input, filename))
+            if filename == 'dataset.csv':
+                os.remove(os.path.join(directory_input, filename))
 
         #saves the new data in the input and input/log directory
         dataset.to_csv(os.path.join(directory_input, 'dataset.csv')
@@ -32,17 +32,25 @@ class DataHandler(object):
     def store_processed_dataset(self, dataset):
 
         directory_input = os.path.join(os.path.dirname(__file__), 'input')
-        #removes all files from the input directory
-        dataset.to_csv(os.path.join(directory_input, 'Pre_Processing.csv'), sep=';', index=None)
+        
+        #removes all processed files from the input directory
+        for file in os.listdir(directory_input):
+            filename = os.fsdecode(file)
+            if filename == 'dataset_processed.csv':
+                os.remove(os.path.join(directory_input, filename))
 
-    def read_network_dataset(self, dataset):
+        #saves the new data in the input and input directory
+        dataset.to_csv(os.path.join(directory_input, 'dataset_processed.csv'),
+                       sep=';', index=None)
+
+    def read_network_dataset(self):
 
         directory_input = os.path.join(os.path.dirname(__file__), 'input')
-        df = pd.read_csv(os.path.join(directory_input, dataset), sep =';')
 
-        return df
+        return pd.read_csv(os.path.join(directory_input, 'dataset.csv'), sep =';')
 
+    def read_processed_dataset(self):
 
+        directory_input = os.path.join(os.path.dirname(__file__), 'input')
 
-
-        
+        return pd.read_csv(os.path.join(directory_input, 'dataset_processed.csv'), sep =';')
